refactor: move all providers into separate file and wrap app with it
